[NFC] Split `common.h` header [2/N]
This PR moves type checking helpers that are only used in a single
test into that test.
