+++ tests/common/common.h
@@ -241,61 +241,6 @@ void check_array_equality(sycl_cts::util::logger& log, ArrT* arr1, ArrT* arr2) {
 }
 
 /**
- * @brief Helper function to see if a type is of the wrong size
- */
-template <typename T>
-bool check_type_min_size(size_t minSize) {
-  return !(sizeof(T) < minSize);
-}
-
-/**
- * @brief Helper function to see if a type is of the wrong sign
- */
-template <typename T>
-bool check_type_sign(bool expected_sign) {
-  return (std::is_signed<T>::value == expected_sign);
-}
-
-#if SYCL_CTS_ENABLE_HALF_TESTS
-/**
- * @brief Helper function to see if sycl::half is of the wrong sign
- */
-template <>
-inline bool check_type_sign<sycl::half>(bool expected_sign) {
-  bool is_signed = sycl::half(1) > sycl::half(-1);
-  return is_signed == expected_sign;
-}
-#endif
-
-/**
- * @brief Helper function to log a failure if a type is of the wrong size or
- * sign
- */
-template <typename T>
-void check_type_min_size_sign_log(size_t minSize, bool expected_sign,
-                                  std::string typeName) {
-  if (!check_type_min_size<T>(minSize)) {
-    FAIL(std::string(
-             "The following host type does not have the correct size: ") +
-         typeName);
-  }
-  if (!check_type_sign<T>(expected_sign)) {
-    FAIL(std::string(
-             "The following host type does not have the correct sign: ") +
-         typeName);
-  }
-}
-
-/**
- * @deprecated Use overload without logger.
- */
-template <typename T>
-void check_type_min_size_sign_log(sycl_cts::util::logger& log, size_t minSize,
-                                  bool expected_sign, std::string typeName) {
-  check_type_min_size_sign_log<T>(minSize, expected_sign, typeName);
-}
-
-/**
  * @brief Verify two values are equal. Deprecated.
  * Use \c value_operations::are_equal instead
  */
