+++ tests/scalars/scalars_interoperability_types.cpp
@@ -20,6 +20,7 @@
 *******************************************************************************/
 
 #include "../common/common.h"
+#include "helpers.h"
 
 #define TEST_NAME scalars_interopability_types
