+++ tests/scalars/scalars_sycl_types.cpp
@@ -20,6 +20,7 @@
 *******************************************************************************/
 
 #include "../common/common.h"
+#include "helpers.h"
 
 #include <array>
